Reimplemented UI with MVC, minor changes

+++ src/gpu/opengl/OpenGl.h
@@ -4,7 +4,6 @@
 
 #pragma once
 
-#include <cassert>
 #include <functional>
 #include <glad/glad.h>
 #include <optional>
@@ -13,7 +12,7 @@ namespace pf {
 
 template<std::invocable<GLuint> Deleter>
 class OpenGlHandle {
-  constexpr static GLuint INVALID_HANDLE = -1;
+  constexpr static GLuint INVALID_HANDLE = static_cast<GLuint>(-1);
 
  public:
   OpenGlHandle()
